firmware_file: Refactored out the FirmwareFormat enum into a new file
