+++ src/lib.rs
@@ -14,6 +14,7 @@ mod bmp_matcher;
 pub mod docs_viewer;
 pub mod elf;
 pub mod error;
+pub mod firmware_file;
 pub mod firmware_selector;
 pub mod flasher;
 pub mod metadata;
